--- hw03/src/main/java/ru/geekbrains/lesson3/ocp/Circle.java
@@ -1,20 +1,0 @@
-package ru.geekbrains.lesson3.ocp;
-
-public class Circle implements Shape {
-
-    private final int radius;
-
-    public int getRadius() {
-        return radius;
-    }
-
-    public Circle(int radius) {
-        this.radius = radius;
-    }
-
-    @Override
-    public double getArea() {
-        return Math.PI * radius * radius;
-    }
-
-}
