--- hw03/src/main/java/ru/geekbrains/lesson3/ocp/Shape.java
@@ -1,6 +1,0 @@
-package ru.geekbrains.lesson3.ocp;
-
-public interface Shape {
-
-    double getArea();
-}
