+++ hw03/src/main/java/ru/geekbrains/lesson3/refueling/Refueling.java
@@ -1,4 +1,6 @@
-package ru.geekbrains.lesson3;
+package ru.geekbrains.lesson3.refueling;
+
+import ru.geekbrains.lesson3.cars.FuelType;
 
 /**
  * Заправочная станция
