--- hw03/src/main/java/ru/geekbrains/lesson3/srp/Point.java
@@ -1,19 +1,0 @@
-package ru.geekbrains.lesson3.srp;
-
-public class Point {
-    private int x;
-    private int y;
-
-    public Point(int x, int y) {
-        this.x = x;
-        this.y = y;
-    }
-
-    public int getY() {
-        return y;
-    }
-
-    public int getX() {
-        return x;
-    }
-}
