+++ hw03/tasks.md
@@ -28,26 +28,28 @@
 конкретный автомобиль путём наследования класса
 «Car», определить число колёс = 3. Провести проверку принципа LSP.
 
-## 6. Создать интерфейс «Заправочная станция», создать метод:
-
+## 6. Создать 
+интерфейс «Заправочная станция», создать метод:
 заправка топливом.
 
 ## 7. Имплементировать
 
 метод интерфейса «Заправочная станция» в
-конкретный класс «Car».
+конкретный класс ~~«Car»~~ **«Заправочная станция»**.
 
 ## 8. Добавить
 
-в интерфейс «Заправочная станция» методы: протирка
+в интерфейс **«Заправочная станция»** методы: протирка
 лобового стекла, протирка фар, протирка зеркал.
 
 ## 9. Имплементировать
 
 все методы интерфейса «Заправочная
-станция» в конкретный класс «Car». Провести проверку
+станция» в конкретный класс ~~«Car»~~ **«Заправочная станция»**.
+Провести проверку
 принципа ISP. Создать дополнительный/е интерфейсы и
-имплементировать их в конкретный класс «Car». Провести
+имплементировать их в конкретный класс ~~«Car»~~ «Заправочная станция».
+Провести
 проверку принципа ISP.
 
 ## 10. Создать
